Remove interactive device configuration & discovery
Let Home Assistant deal with device discovery, and let the Almond
UI deal with configuring skills. This kills the "interactive" mode
of configuring devices, but that's only used by Matrix so we don't
need to worry about it.
