+++ lib/dialogue-agent/almond.js
@@ -178,10 +178,6 @@ module.exports = class Almond extends events.EventEmitter {
         return this._engine.thingpedia;
     }
 
-    get discovery() {
-        return this._engine.discovery;
-    }
-
     get remote() {
         return this._engine.remote;
     }
@@ -196,10 +192,6 @@ module.exports = class Almond extends events.EventEmitter {
 
     notifyError() {
         return this._dispatcher.dispatchNotifyError.apply(this._dispatcher, arguments);
-    }
-
-    interactiveConfigure() {
-        return this._dispatcher.dispatchInteractiveConfigure.apply(this._dispatcher, arguments);
     }
 
     expect(expecting) {
