+++ lib/dialogue-agent/dialogue_loop.js
@@ -13,8 +13,9 @@ const ThingTalk = require('thingtalk');
 const Ast = ThingTalk.Ast;
 
 const { chooseDevice, concretizeValue } = require('./entity-linking');
+const { showNotification, showError } = require('./notifications');
 
-const { handleLegacyAPICall, legacyDialogueHandler } = require('./policies/legacy');
+const { legacyDialogueHandler } = require('./policies/legacy');
 
 const Helpers = require('./helpers');
 
@@ -277,6 +278,19 @@ async function handleUserInput(dlg, state, input) {
     return newStyleDialogueHandler(dlg, state, input.intent);
 }
 
+async function handleAPICall(dlg, call, lastApp) {
+    let value;
+    if (call.isNotification) {
+        value = await showNotification(dlg, call.appId, call.icon, call.outputType, call.outputValue, lastApp);
+        lastApp = call.appId;
+    } else if (call.isError) {
+        value = await showError(dlg, call.appId, call.icon, call.error, lastApp);
+        lastApp = call.appId;
+    }
+
+    return [value, lastApp];
+}
+
 module.exports = async function loop(dlg, showWelcome) {
     let lastApp = undefined, currentDialogueState = null, next = undefined;
     for (;;) {
@@ -296,7 +310,7 @@ module.exports = async function loop(dlg, showWelcome) {
                 lastApp = undefined;
                 currentDialogueState = await handleUserInput(dlg, currentDialogueState, current);
             } else {
-                [value, lastApp] = await handleLegacyAPICall(dlg, current, lastApp);
+                [value, lastApp] = await handleAPICall(dlg, current, lastApp);
                 currentDialogueState = null;
             }
