+++ lib/dialogue-agent/dialogue_queue.js
@@ -79,17 +79,4 @@ class Error extends QueueItem {
 Error.prototype.isError = true;
 QueueItem.Error = Error;
 
-class InteractiveConfigure extends QueueItem {
-    constructor(kind) {
-        super();
-        this.kind = kind;
-    }
-
-    toString() {
-        return `InteractiveConfigure(${this.kind})`;
-    }
-}
-InteractiveConfigure.prototype.isInteractiveConfigure = true;
-QueueItem.InteractiveConfigure = InteractiveConfigure;
-
 module.exports = QueueItem;
