+++ lib/dialogue-agent/dispatcher.js
@@ -319,10 +319,6 @@ module.exports = class Dispatcher {
         return this._notifyQueue.hasWaiter();
     }
 
-    dispatchInteractiveConfigure(kind) {
-        let item = new QueueItem.InteractiveConfigure(kind);
-        return this._pushQueueItem(item);
-    }
     dispatchNotify(appId, icon, outputType, outputValue) {
         let item = new QueueItem.Notification(appId, icon, outputType, outputValue);
         return this._pushQueueItem(item);
