+++ lib/dialogue-agent/notifications.js
@@ -9,7 +9,7 @@
 // See COPYING for details
 "use strict";
 
-const Helpers = require('../helpers');
+const Helpers = require('./helpers');
 
 async function showNotification(dlg, appId, icon, outputType, outputValue, lastApp) {
     let app;
