+++ lib/dialogue-agent/policies/legacy.js
@@ -9,34 +9,12 @@
 // See COPYING for details
 "use strict";
 
-const discoveryDialog = require('../legacy-dialogs/discovery');
-const configDialog = require('../legacy-dialogs/config');
 const { fallback, getExamples } = require('../legacy-dialogs/fallback');
-const { showNotification, showError } = require('../legacy-dialogs/notifications');
 
 const Semantic = require('../semantic');
 const ValueCategory = Semantic.ValueCategory;
 const Helpers = require('../helpers');
 const { CancellationError } = require('../errors');
-
-async function handleAPICall(dlg, call, lastApp) {
-    let value;
-    if (call.isNotification) {
-        value = await showNotification(dlg, call.appId, call.icon, call.outputType, call.outputValue, lastApp);
-        lastApp = call.appId;
-    } else if (call.isError) {
-        value = await showError(dlg, call.appId, call.icon, call.error, lastApp);
-        lastApp = call.appId;
-    } else if (call.isInteractiveConfigure) {
-        lastApp = undefined;
-        if (call.kind !== null)
-            value = await configDialog(dlg, call.kind);
-        else
-            value = await discoveryDialog(dlg);
-    }
-
-    return [value, lastApp];
-}
 
 async function handleInput(dlg, intent, confident) {
     if (intent.isFailed)
@@ -150,16 +128,6 @@ async function legacyDialogueHandler(dlg, input) {
 }
 
 
-async function handleLegacyAPICall(dlg, input, lastApp) {
-    dlg.setBeforeInput(legacyHandleGeneric);
-    try {
-        return await handleAPICall(dlg, input, lastApp);
-    } finally {
-        dlg.setBeforeInput(null);
-    }
-}
-
 module.exports = {
-    handleLegacyAPICall,
     legacyDialogueHandler,
 };
